Remove unnecessary cruft.
Here I am stripping away as much as possible. Someday I'll get to
writing, but first I need a clean slate.

Tonight it's a Boulevardier—an ounce each of Bourbon, Sweet Vermouth,
and Campari. Shake it with ice, strain, and pour it into a glass with a
peel of orange and a single ice cube. Give it five minutes before you
take a sip.